Cancel a reader pause, not merely raise inside it
`_paused_reader` resumes in a bare `finally:`, so a `CancelledError` — a
`BaseException` — still undoes the pause. Nothing proved it. The only
test driving both exits parametrized `body_raises` over a synchronously
raised `ValueError`, which an `except Exception:` guard catches just as
well, so narrowing the guard would skip the resume on cancellation and
leave the whole suite green.

Add a test that parks a task on an await inside the block and cancels it
from outside, asserting both that the `CancelledError` reaches the
awaiter and that the transport is resumed exactly once. Under the
narrowed guard it fails on `resume_calls == 0`, while every other test —
including the `ValueError` case — still passes.

Adding it would have pushed the file past the 400-line cap, so the two
`_BlockingModeGuard` property tests move to their own module, along the
seam the module docstring already named. The new file shifts the maturin
wheel manifest, so its snapshot is regenerated.
